Updated MotionSystem to support Clock hand motion CLocks now showing correct time.

+++ ClockHandMotionComponent.cpp
@@ -18,9 +18,9 @@ MotionComponent * ClockHandMotionComponent::clone() const
 }
 
 //---------------------------------------------------------------------------------------
-void ClockHandMotionComponent::update(
+void ClockHandMotionComponent::update (
 	GameObject * gameObject,
-	double elapsedTime
+	float elapsedTime
 ) {
 
 	MeshId meshId = gameObject->graphics->mesh->meshId;
